Add lore
Also fix small bug.

I was going to add this to a README after initially thinking having it
embedded in the command was unnecessary, but since it uses your record
and the command seems to be working well I'm adding it, probably no
harm. See comments in code on separating all non-fishing action to a
separate command if it's a problem.