Project overhaul
Most changes concern trying to make this simple game operate as simply
as possible.

Loading the command into memory should be light, it happens every
command (I think). Reduce objects needing to be loaded into memory
every time. Keep the script small size (not sure how much this
matters). While static maps are included at all times, they are loaded
as functions now instead of large objects except for the current map,
which gets loaded. (I think this should be faster.)

Objects should be used more instead of arrays. The inventory/history
data object on the player is pretty stupid, it should've probably been
one keyed object from the start. (Would make `collection` not show your
history of catches, however). However, with a cap of 200 on inventory,
and O(types of fish) on iterating history, it's probably entirely
negligible.

The call stack from `$$ gofish -> Nothing...` should be as bare as
possible, though this doesn't matter that much and is micro
optimization-y. Still good to think about, this is ~93% of the command.
Anything outside this call stack can do whatever it wants.

And some other small stuff. Basically just need to be careful what is
happening from beginning to "miss", including how much memory the
script is taking (as in when loaded not filesize). Maybe entirely
pointless but I'd rather do it right.

This also just makes the project structure better and closes #4.